PDDF: Add PddfComponent type to track programmables
In SONiC devices, many components may have upgradable firmware, beyond
just the SONiC image itself. We can track these versions and upgrades
through a new PddfComponent type. The components will be defined in the
pddf-device.json with a method to read model, version, and revision, as
well as update the firmware.

Add tpm2-tools to the SONiC image in order to manage TPM devices using
PddfComponent.